requirements.txt
asgiref==3.2.5
astroid==2.3.3
autopep8==1.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
google-api-core==1.16.0
google-auth==1.11.3
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
wrapt==1.11.2
